+++ src/ampworks/gitt/__init__.py
@@ -1,6 +1,4 @@
 """
-GITT
-----
 Tools to analyze Galvanostatic Intermittent Titration Technique (GITT) data.
 Includes functions to extract diffusion coefficients and equilibrium potentials
 from experimental measurements.